removed unused lib, move BottomNavigation.kt from dump package to navigation package
