+++ app/src/main/java/com/cyberwalker/fashionstore/home/HomeScreen.kt
@@ -15,20 +15,20 @@
  */
 package com.cyberwalker.fashionstore.home
 
-import android.util.Log
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.layout
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.semantics.contentDescription
@@ -37,19 +37,14 @@ import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
-import androidx.navigation.Navigation
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import coil.compose.AsyncImagePainter.State.Empty.painter
 import com.cyberwalker.fashionstore.R
-import com.cyberwalker.fashionstore.detail.DetailScreen
-import com.cyberwalker.fashionstore.dump.BottomNav
 import com.cyberwalker.fashionstore.dump.vertical
-import com.cyberwalker.fashionstore.navigation.*
+import com.cyberwalker.fashionstore.navigation.BottomNav
+import com.cyberwalker.fashionstore.navigation.DetailsScreen
+import com.cyberwalker.fashionstore.navigation.HomeNavGraph
 import com.cyberwalker.fashionstore.ui.theme.*
-import com.google.accompanist.navigation.animation.rememberAnimatedNavController
-import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.launch
 
 private const val TAG = "HomeScreen"
 @Composable
