+++ app/src/main/java/com/cyberwalker/fashionstore/search/SearchScreen.kt
@@ -7,11 +7,9 @@ import androidx.compose.material.Text
 import androidx.compose.material.rememberScaffoldState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
-import com.cyberwalker.fashionstore.dump.BottomNav
 import com.cyberwalker.fashionstore.home.HomeScreenActions
-import com.cyberwalker.fashionstore.home.HomeViewModel
+import com.cyberwalker.fashionstore.navigation.BottomNav
 
 @Composable
 fun SearchScreen(
